Fix build on macOS.
This changeset consists of two changes:

The declaration of the pyInitLastState global variable is moved from
pydevsup.h to setup.c. This is needed to avoid an "duplicate symbols"
error when linking. As this variable is only used in setup.c anyway,
there is no downside in moving its declaration there.

The log example application is now only built on Linux. As it depends
on the inotify API, which is Linux-specific, it cannot be reasonably
built on other platforms. As it is not needed for the pyDevSup to work,
it is OK to built it without this application.