Updated RULES_PY to use @$(INSTALL). This requires include $(CONFIG)/CONFIG_BASE for the definition of the INSTALL macro.
epicsThreadExitMain is deprecated by recent EPICS base versions (and resolves to cantProceed). Being unsure as to what best to do about this, I commented it out in devSupMain.cpp.

Python 2.7 has been obsoleted and won't install on the runner without taking special measures.
Being unsure what to do about this, I have commented out builds that invoke 2.7.

Github Windows-latest (i.e. Windows server 2022) is installed with VS2022 (not VS2019).
Use of the debug build configuration will seek to link with the Python debug libraries that are not normally installed. I have a solution to this problem but we should debate how to proceed.

Github Windows server 2019 is, I think, installed with VS2019, this should build OK with default configuration.